Update Aurora.py
Noticed several issues...  double drawing when colored tiles enabled so commented out the double call around line 458... so i tweaked the offset of the tiles. This fix probably only works for my screen resolution. Try tweaking the offset of the tiles on line 411 and 412 to about -2.6 and -2.2   which was i think was -.5 and -.5 before w other monitor and screen resolution... hmm